Nouveau répertoire "tutorial" et suppression de 2 slides pour l'enquête
Chaque tuto devient un composant
