+++ source/sites/publicodes/App.tsx
@@ -48,7 +48,7 @@ const PetrogazLandingLazy = React.lazy(() => import('./pages/PetrogazLanding'))
 const ModelLazy = React.lazy(() => import('./Model'))
 const PersonasLazy = React.lazy(() => import('./Personas'))
 const DocumentationLazy = React.lazy(() => import('./pages/Documentation'))
-const TutorialLazy = React.lazy(() => import('./Tutorial'))
+const TutorialLazy = React.lazy(() => import('./tutorial/Tutorial'))
 const GroupSwitchLazy = React.lazy(() => import('./conference/GroupSwitch'))
 const ContributionLazy = React.lazy(() => import('./Contribution'))
 const ConferenceLazy = React.lazy(() => import('./conference/Conference'))
